WIP: Restructure giftcards analogue to discounts
